Handle zeros
Ran this on the sample input and it didn't match up. Hunted for a bit and noticed it wasn't halting at the right number. 
Dug a bit, realized if zero is called, multiplying by -1 doesn't "mark the board".

Thanks for posting your code; I tried solving using lists of lists of lists and... well, it got messy.  
cheers.